# app/docs/swagger.yaml
definitions:
...
  alertModels.AlertLogData:
    properties:
      alerts:
        items:
          properties:
            annotations:
              properties:
                reason:
                  type: string
              type: object
            endsAt:
              type: string
            labels:
              properties:
                alert_id:
                  type: string
                context:
                  type: string
                data_id:
                  type: string
                device_id:
                  type: string
                group_id:
                  type: string
                message:
                  type: string
                template_id:
                  type: string
              type: object
            startsAt:
              type: string
          type: object
        type: array
      commonLabels:
        additionalProperties:
          type: string
        type: object
      gatewayId:
        type: integer
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    type: object
...
  alertModels.SysAlertLogListReq:
    properties:
      alert_id:
        description: 告警id
        type: integer
      gateway_id:
        description: 网关id
        type: integer
      isAsc:
        description: 排序规则  降序desc   asc升序
        type: string
      message:
        type: string
      orderBy:
        description: 排序字段
        type: string
      pageNum:
        default: 1
        description: 第几页
        type: integer
      pageSize:
        default: 10000
        description: 数量
        type: integer
    type: object
  baize.List:
    properties:
      data:
        items:
          type: string
        type: array
    type: object
...
  craftRouteModels.SysCraftRouteRequest:
    properties:
      datasetId:
        description: 出库id
        example: "0"
        type: string
      remark:
        description: 备注
        type: string
      route_code:
        description: 工艺路线编号
        type: string
      route_desc:
        description: 工艺路线说明
        type: string
      route_name:
        description: 工艺路线名称
        type: string
      status:
        description: 操作状态（0正常 1异常）
        type: boolean
    required:
    - route_code
    - route_name
    type: object
...
  craftRouteModels.SysProductScheduleReq:
    properties:
      month:
        type: string
      year:
        type: string
    type: object
...
  monitorModels.JobLog:
    properties:
      costTime:
        type: integer
      createTime:
        type: string
      exceptionInfo:
        type: string
      invokeTarget:
        type: string
      jobId:
        example: "0"
        type: string
      jobLogId:
        example: "0"
        type: string
      jobName:
        type: string
      jobParams:
        $ref: '#/definitions/baize.List'
      status:
        type: string
    type: object
...
  response.ListData:
    properties:
      rows: {}
      total:
        type: integer
    type: object
  response.ResCode:
    enum:
    - 200
    - 401
    - 403
    - 412
    - 500
    - 600
    type: integer
    x-enum-comments:
      Error: 系统异常
      Forbidden: 没有权限
      Parameter: 参数错误
      Success: 成功
      Unauthorized: token失效
      Waring: 详情看msg
    x-enum-varnames:
    - Success
    - Unauthorized
    - Forbidden
    - Parameter
    - Error
    - Waring
  response.ResponseData:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/response.ResCode'
        description: 相应状态码
      data:
        description: 数据
      msg:
        description: 提示信息
        type: string
    type: object
...
info:
  contact:
    email: zhao_402295440@126.com
    name: danny
  description: baize接口文档
  title: baize
  version: 2.0.x
paths:
...
  /alert/log/list:
    get:
      description: 告警数据列表
      parameters:
      - description: 助理列表参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/alertModels.SysAlertLogListReq'
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 告警数据列表
      tags:
      - 告警管理/告警数据管理
...
  /api/alert/log/v1/export:
    post:
      description: 导入告警数据
      parameters:
      - description: 助理列表参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/alertModels.AlertLogData'
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 导入告警数据
      tags:
      - 告警管理/告警数据管理
...
  /craft/route/list:
    get:
      description: 读取工艺列表
      parameters:
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - description: 工艺路线编号
        in: query
        name: route_code
        type: string
      - description: 工艺路线名称
        in: query
        name: route_name
        type: string
      - description: 操作状态（0正常 1异常）
        in: query
        name: status
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 读取工艺列表
      tags:
      - 工艺管理
...
  /craft/route/remove/{craft_route_id}:
    delete:
      description: 移除工艺
      parameters:
      - description: craft_route_id
        in: path
        name: craft_route_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 设置分组成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 移除工艺
      tags:
      - 工艺管理
  /craft/route/schedule/month/list:
    get:
      description: 月调度列表
      parameters:
      - description: 组成工序列表参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/craftRouteModels.SysProductScheduleReq'
      produces:
      - application/json
      responses:
        "200":
          description: 设置分组成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 月调度列表
      tags:
      - 工艺管理/调度管理
  /craft/route/set:
    post:
      description: 设置工艺路线
      parameters:
      - description: 设备分组参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/craftRouteModels.SysCraftRouteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 设置分组成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 设置工艺路线
      tags:
      - 工艺管理
...
  /monitor/job/log/list:
    get:
      description: 查询定时任务日志列表
      parameters:
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - in: query
        name: jobId
        type: integer
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListData'
                  - properties:
                      Rows:
                        items:
                          $ref: '#/definitions/monitorModels.JobLog'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 查询定时任务日志列表
      tags:
      - 定时任务
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"